Improve and move steamid and steamprofile detect code.
