+++ Source/IdleMaster/SteamProfile.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using System.Net;
+using System.Text;
+using System.Xml;
+using IdleMaster.Properties;
+
+namespace IdleMaster
+{
+  internal static class SteamProfile
+  {
+    internal static string GetSteamId()
+    {
+      var steamid = WebUtility.UrlDecode(Settings.Default.steamLogin);
+      var index = steamid.IndexOfAny(new[] { '|' }, 0);
+      return index != -1 ? steamid.Remove(index) : steamid;
+    }
+
+    internal static string GetSteamUrl()
+    {
+      return "http://steamcommunity.com/profiles/" + GetSteamId();
+    }
+
+    public static string GetSignedAs()
+    {
+      var steamid = GetSteamUrl();
+      var userName = "User " + GetSteamId();
+      try
+      {
+        var xmlRaw = new WebClient() { Encoding = Encoding.UTF8 }.DownloadString(string.Format("{0}/?xml=1", steamid));
+        var xml = new XmlDocument();
+        xml.LoadXml(xmlRaw);
+        var nameNode = xml.SelectSingleNode("//steamID");
+        if (nameNode != null)
+          userName = WebUtility.HtmlDecode(nameNode.InnerText);
+      }
+      catch (Exception ex)
+      {
+        Console.WriteLine(ex.Message);
+        Logger.Exception(ex, "frmMain -> GetSignedAs, for steamid = " + steamid);
+      }
+      return "Signed in as " + userName;
+    }
+  }
+}
